Task 4.2D/requirements.txt
joblib>=1.3.0
numpy>=1.23.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
azure-ai-ml>=1.15.0
azure-identity>=1.16.0

